added configuration output path and checks for null on output port

+++ jsymbolic_utilities.py
@@ -44,7 +44,7 @@ def execute(cmdArray, workingDir):
 
 
 def copy_when_exists(src, dst):
-    if os.path.isfile(src):
+    if os.path.isfile(src) and dst is not None:
         shutil.copyfile(src, dst)
